requirements.txt
# Core Processing
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
tqdm>=4.65.0
requests>=2.31.0

# Modern UI Framework (Glassmorphism)
PySide6>=6.6.0
PySide6-Fluent-Widgets>=1.4.0
pyqt-frameless-window>=0.3.0
darkdetect>=0.8.0
